refractor index and App
move logics in index to App, deleted a little bit unused codes.
Literally not changing anything in other files, just removing header and outer wrapper.
